CORE-7943 move CommonLedgerIntegrationTest away to not disturb persistence tests
