+++ settings.gradle
@@ -392,6 +392,7 @@ include 'testing:flow:external-events'
 include 'testing:flow:flow-utilities'
 include 'testing:layered-property-map-testkit'
 include 'testing:ledger:ledger-common-base-test'
+include 'testing:ledger:ledger-common-base-integrationtest'
 include 'testing:ledger:ledger-common-testkit'
 include 'testing:ledger:ledger-common-empty-app'
 include 'testing:ledger:ledger-consensual-base-test'
