+++ testing/ledger/ledger-common-base-integrationtest/src/main/java/net/corda/ledger/common/integrationtest/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.ledger.common.integrationtest;
+
+import org.osgi.annotation.bundle.Export;
