+++ testing/ledger/ledger-common-testkit/build.gradle
@@ -6,7 +6,6 @@ description 'Corda ledger common test objects'
 
 dependencies {
     compileOnly 'org.osgi:osgi.annotation'
-    compileOnly 'org.osgi:osgi.core'
 
     implementation platform("net.corda:corda-api:$cordaApiVersion")
     api 'net.corda:corda-ledger-common'
@@ -16,12 +15,4 @@ dependencies {
     api project(':components:ledger:ledger-common-flow-api')
     api project(':libs:application:application-impl')
     api project(':libs:platform-info')
-
-    api project(':testing:sandboxes')
-    api project(':testing:sandboxes-testkit')
-    implementation project(':components:flow:flow-service')
-    implementation project(':libs:serialization:serialization-checkpoint-api')
-    implementation project(':libs:serialization:serialization-internal')
-    implementation project(':libs:serialization:serialization-amqp')
-    implementation "org.osgi:org.osgi.test.junit5:$osgiTestJunit5Version"
 }
